Safari-609.1.11.1/LayoutTests/svg/animations/svgenum-animation-5-expected.txt

SVG 1.1 dynamic animation tests

Test MorphologyOperatorType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS morphology.operator.animVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.animVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.animVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_ERODE
PASS morphology.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.animVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_ERODE
PASS morphology.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.animVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphology.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS successfullyParsed is true

TEST COMPLETE
